+++ rtsim/src/gen/mod.rs
@@ -289,9 +289,19 @@ impl Data {
         }
         // Spawn one monster Gigasfrost into the world
         // Try a few times to find a location that's not underwater
-        if let Some((wpos, chunk)) = (0..10)
+        if let Some((wpos, chunk)) = (0..100)
             .map(|_| world.sim().get_size().map(|sz| rng.gen_range(0..sz as i32)))
-            .find_map(|pos| Some((pos, world.sim().get(pos).filter(|c| !c.is_underwater())?)))
+            .find_map(|pos| {
+                Some((
+                    pos,
+                    world
+                        .sim()
+                        .get(pos)
+                // This is currently a workaround to force Frost Gigas spawning in cold areas
+                // TODO: Once more Gigas are implemented remove this
+                        .filter(|c| !c.is_underwater() && c.temp < CONFIG.snow_temp)?,
+                ))
+            })
             .map(|(pos, chunk)| {
                 let wpos2d = pos.cpos_to_wpos_center();
                 (